pyproject.toml
[project]
name = "binance-um-price-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "websocket-client>=1.8.0",
]
